+++ app/components/primer/avatar_stack_component.rb
@@ -15,8 +15,8 @@ module Primer
     BODY_TAG_OPTIONS = TAG_OPTIONS = [DEFAULT_TAG, :span].freeze
     # Required list of stacked avatars.
     #
-    # @param kwargs [Hash] The same arguments as <%= link_to_component(Primer::AvatarComponent) %>.
-    renders_many :avatars, Primer::AvatarComponent
+    # @param kwargs [Hash] The same arguments as <%= link_to_component(Primer::Beta::Avatar) %>.
+    renders_many :avatars, "Primer::Beta::Avatar"
 
     # @example Default
     #   <%= render(Primer::AvatarStackComponent.new) do |c| %>